refactor: move server code to a package
... to make it importable in other projects. For example to use it with
the `InProcessClient` `from mark3labs/mcp-go`
